+++ src/main/java/com/dionialves/snapdogdelivery/domain/admin/productorder/ProductOrderRepository.java
@@ -1,0 +1,8 @@
+package com.dionialves.snapdogdelivery.domain.admin.productorder;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductOrderRepository extends JpaRepository<ProductOrder, Long> {
+
+    boolean existsByProductId(Long productId);
+}